+++ packages/polaris-viz/src/components/VerticalBarChart/components/VerticalBar/tests/VerticalBar.test.tsx
@@ -45,15 +45,15 @@ describe('<VerticalBar/>', () => {
       expect(bar).toContainReactComponent('path', {
         // eslint-disable-next-line id-length
         d: `
-  M0,0
-  h100
-  a0,0 0 0 1 0,0
-  v1000
+  M3,0
+  h94
+  a3,3 0 0 1 3,3
+  v997
   a0,0 0 0 1 -0,0
   h-100
   a0,0 0 0 1 -0,-0
-  v-1000
-  a0,0 0 0 1 0,-0
+  v-997
+  a3,3 0 0 1 3,-3
   Z
 `,
       });